Modify RateLimiter to use topic as String instead of Enum
In order to provide the greatest level of isolation between system updates,
having the possibility for external systems to define new topics and config updates,
the email topic has become a String instead of an Enum.

Now, the only reason for a new deployment of a service implementing the rate-limiter
would be to change its actual implementation, not its supported topics.

Since it's no longer possible to guarantee that all message topics processed have a config,
the RateLimiterProvider behavior has changed slightly and in cases where there's no defined config,
the behavior is set to 'prohibited' and an error log is produced.

By using the new test extension 'LogErrorVerifierExtension', it's possible to guarantee that a test
does not try to send an email using unconfigured topics.